requirements.txt
streamlit>=1.30
google-genai>=1.0
pandas>=2.0
python-dotenv>=1.0
Pillow>=10.0
pydantic>=2.0
